backend/requirements.txt
annotated-doc
annotated-types
anyio
bcrypt
certifi
click
colorama
dnspython
ecdsa
email-validator
fastapi
h11
httpcore
httpx
idna
iniconfig
motor
packaging
passlib
pydantic
pydantic-settings
pydantic_core
pymongo
pytest
pytest-asyncio
python-dotenv
python-jose
python-multipart
rsa
six
starlette
typing-inspection
typing_extensions
uvicorn
pluggy
pip
Pygments
